requirement.txt
Flask
pdfplumber
docx2txt
python-docx
openai
gunicorn